+++ src/Entities/Reporting/BaseSummary.php
@@ -1,0 +1,36 @@
+<?php
+
+
+namespace GlobalPayments\Api\Entities\Reporting;
+
+
+abstract class BaseSummary
+{
+    /**
+     * @var integer
+     */
+    public $amount;
+    /**
+     * @var string
+     */
+    public $currency;
+    /**
+     * @var string
+     */
+    public $merchantId;
+
+    /**
+     * @var string
+     */
+    public $merchantHierarchy;
+
+    /**
+     * @var string
+     */
+    public $merchantName;
+
+    /**
+     * @var string
+     */
+    public $merchantDbaName;
+}