+++ test/Integration/Gateways/GpApiConnector/GpApi3DSecureTest.php
@@ -9,6 +9,7 @@ use GlobalPayments\Api\Entities\Enums\AuthenticationSource;
 use GlobalPayments\Api\Entities\Enums\ChallengeWindowSize;
 use GlobalPayments\Api\Entities\Enums\Channel;
 use GlobalPayments\Api\Entities\Enums\ColorDepth;
+use GlobalPayments\Api\Entities\Enums\ExemptStatus;
 use GlobalPayments\Api\Entities\Enums\ManualEntryMethod;
 use GlobalPayments\Api\Entities\Enums\MethodUrlCompletion;
 use GlobalPayments\Api\Entities\Enums\OrderTransactionType;
@@ -879,5 +880,22 @@ class GpApi3DSecureTest extends TestCase
         $this->assertNotNull($response);
         $this->assertEquals('SUCCESS', $response->responseCode);
         $this->assertEquals(TransactionStatus::CAPTURED, $response->responseMessage);
+    }
+
+    public function testExemptionSaleTransaction()
+    {
+        $this->card->number = GpApi3DSTestCards::CARD_CHALLENGE_REQUIRED_V2_2;
+
+        $threeDS = new ThreeDSecure();
+        $threeDS->exemptStatus = ExemptStatus::LOW_VALUE;
+        $this->card->threeDSecure = $threeDS;
+        $response = $this->card->charge($this->amount)
+            ->withCurrency($this->currency)
+            ->execute();
+
+        $this->assertNotNull($response);
+        $this->assertEquals('SUCCESS', $response->responseCode);
+        $this->assertEquals(TransactionStatus::CAPTURED, $response->responseMessage);
+
     }
 }